Backend/requirements.txt
# FarmersHub API Requirements
# Core FastAPI and web framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database dependencies
sqlalchemy==2.0.23
sqlite3  # Built-in with Python
alembic==1.12.1

# AI and Machine Learning dependencies
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.3
joblib==1.3.2
transformers==4.35.2
torch==2.1.1
tensorflow==2.15.0

# Hugging Face dependencies
huggingface-hub==0.19.4
datasets==2.14.6

# HTTP and API dependencies
requests==2.31.0
httpx==0.25.2
aiohttp==3.9.1

# Image processing dependencies
Pillow==10.1.0
opencv-python==4.8.1.78

# Data processing and analysis
scipy==1.11.4
matplotlib==3.8.2
seaborn==0.13.0

# Caching and performance
redis==5.0.1
cachetools==5.3.2

# Configuration and environment
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Logging and monitoring
loguru==0.7.2
structlog==23.2.0

# Security
cryptography==41.0.8
bcrypt==4.1.2

# File handling
aiofiles==23.2.1
python-magic==0.4.27

# Date and time
python-dateutil==2.8.2
pytz==2023.3

# JSON and serialization
orjson==3.9.10
ujson==5.8.0

# Testing dependencies (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development dependencies (optional)
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Documentation dependencies (optional)
mkdocs==1.5.3
mkdocs-material==9.4.8

# Production dependencies
gunicorn==21.2.0
psycopg2-binary==2.9.9  # For PostgreSQL support
celery==5.3.4  # For background tasks
flower==2.0.1  # For Celery monitoring

# Mobile and PWA dependencies
pywebpush==1.14.0  # For push notifications
pyjwt==2.8.0  # For JWT tokens

# Additional utilities
click==8.1.7
rich==13.7.0
typer==0.9.0
tqdm==4.66.1

# Supabase dependencies
supabase==2.3.4
postgrest==0.13.2

# Weather API dependencies
pyowm==3.3.0  # Alternative weather API

# Market data dependencies
yfinance==0.2.28  # For market data
beautifulsoup4==4.12.2  # For web scraping
lxml==4.9.3

# Email dependencies
fastapi-mail==1.4.1
jinja2==3.1.2

# Background task dependencies
apscheduler==3.10.4

# Monitoring and metrics
prometheus-client==0.19.0
